+++ composer.json
@@ -14,11 +14,13 @@
         }
     ],
     "require": {
-        "guzzlehttp/guzzle": "~7.4",
         "php": ">=7.4",
-        "ext-json": "*"
+        "ext-json": "*",
+        "psr/http-message": "^1.0",
+        "psr/http-client": "^1.0",
+        "psr/http-factory": "^1.0"
     },
     "autoload": {
-        "psr-0": {"Foxy\\FoxyClient": "src/"}
+        "psr-4": {"Foxy\\FoxyClient\\": "src/"}
     }
 }
